CLOUD-33: Handle too old resource version
This is probably caused by bug in kubernetes client -
when requesting broad event stream (like all ingresses with specific label),
the client picks up resource version from first event, and then it fails
on reconnect because the resource version is too old.

The fix handles the exception, and will use the recommended resource
version for reconnection.